Allow same tracking ID to track by offset and timestamp
Before this commit, a tracking ID could store either an offset
or a timestamp (within the same Pending collection, i.e. no flushing in
between).

After this commit, a tracking ID can store both an offset and a
timestamp.

Although there is probably no use case for this right now, it will cause
less confusion in future if a client happens to track by both timestamp
and offset for the same tracking ID.